+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Version 1.5.1 (not released yet)
 - add icon for EAX Effect objects in the level editor (icon provided by Goober)
 - add more textures from base game to the level editor texture browser
 - fix packing a level that contains a particle emitter with a default texture in the level editor
+- fix path node connections sometimes being rendered incorrectly in the level editor
 
 Version 1.5.0 (released 2020-09-05)
 -----------------------------------